camera/capture_video_cancel-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset='utf-8'>
    <title>HTML Media Capture Test: user denial of captured video leading to no capture</title>
    <link rel='author' title='Intel' href='http://www.intel.com'>
    <link rel='help' href='http://www.w3.org/TR/html-media-capture/#security'>
    <link rel='help' href='http://www.w3.org/TR/html5/forms.html#file-upload-state-(type=file)'>
    <meta name='flags' content='interact'>
    <script src='../support/testharness.js'></script>
    <script src='../support/testharnessreport.js'></script>
  </head>
  <body>
    <p>This test validates that the user agent must not save the captured video to any data storage if the user cancel the capture.</p>

    <p>Test steps:</p>
    <ol>
      <li>Click the button. If prompted for permission, please allow.</li>
      <li>Capture a video, then cancel captured video. (Make sure do it in 10 seconds).</li>
    </ol>

    <form>
      <input type="file" accept="video/*" capture>
    </form>

    <div id='log'></div>

    <script>
      var timeout;
      setup({explicit_done: true, explicit_timeout: true});

      var input = document.querySelector('input[type=file]');

      input.onchange = function() {
        clearTimeout(timeout);
        test(function() {
          assert_equals(input.files.length, 0, 'no data is captured');
        }, document.title);

        done();
      }

      input.onclick = function() {
        timeout = setTimeout(function () {
            test(function() {
              assert_equals(input.files.length, 0, 'no data is captured');
            }, document.title);

            done();
        }, 10000);
      }
    </script>
  </body>
</html>
